main.py
import kagglehub as kh
import os
import pandas as pd
from PIL import Image
import numpy as np


path = kh.dataset_download("masoudnickparvar/brain-tumor-mri-dataset")
print("Path to dataset files:", path)

