requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.19.1
fsspec==2025.10.0
idna==3.11
jinja2==3.1.6
kagglehub==0.3.13
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pyyaml==6.0.3
requests==2.32.5
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-extensions==4.15.0
urllib3==2.5.0
